Convert ACLChecklist::accessList to CbcPointer
Also fixes a dangerous bug in Checklist::changeAcl() where
ACLChecklist::accessList was dereferenced before the caller got
to re-reference the output. This should not have been causing
problems because the access list should have had at least one
external reference in the config structures. However, that may
not be always existing during or after a reconfigure.

changeAcl() will now ensure the reference stays valid in a
Pointer until the caller has completed its re-reference
(or not).