Move fakeReadBuf_ into IdleConnList::push
This data buffer is not used by Squid, simply provided as a place
in memory for I/O to read() packet data into - where it will be
ignored.

As such there is no need for each idle connection to provide its
own buffer, they can all share the same memory that gets
overwritten. IdleConnList was already sharing the buffer between
all connections to the same server in its theList member. We
extend that to all connections stored by any IdleConnList object.

This reduces the memory overheads of idle connections from up to
several MB down to a fixed 4KB.